misc: Move all gobjects to a separate tdlib folder
This keeps them more separated from the widgets.
