+++ src/main.rs
@@ -2,14 +2,13 @@ mod application;
 #[rustfmt::skip]
 #[allow(clippy::all)]
 mod config;
-mod country_info;
-mod country_list;
 mod expressions;
 mod login;
 mod phone_number_input;
 mod preferences_window;
 mod session;
 mod session_manager;
+mod tdlib;
 mod utils;
 mod window;
