+++ src/tdlib/mod.rs
@@ -1,0 +1,44 @@
+mod action;
+mod action_list;
+mod avatar;
+mod basic_group;
+mod basic_group_list;
+mod chat;
+mod chat_list;
+mod country_info;
+mod country_list;
+mod history;
+mod item;
+mod message;
+mod message_forward_info;
+mod secret_chat;
+mod secret_chat_list;
+mod sponsored_message;
+mod supergroup;
+mod supergroup_list;
+mod user;
+mod user_list;
+
+pub(crate) use self::action::ChatAction;
+pub(crate) use self::action_list::ChatActionList;
+pub(crate) use self::avatar::Avatar;
+pub(crate) use self::basic_group::BasicGroup;
+pub(crate) use self::basic_group_list::BasicGroupList;
+pub(crate) use self::chat::{
+    BoxedChatMemberStatus, BoxedChatNotificationSettings, BoxedChatPermissions, BoxedDraftMessage,
+    BoxedMessageContent, Chat, ChatType,
+};
+pub(crate) use self::chat_list::ChatList;
+pub(crate) use self::country_info::CountryInfo;
+pub(crate) use self::country_list::CountryList;
+use self::history::History;
+pub(crate) use self::item::{Item, ItemType};
+pub(crate) use self::message::{Message, MessageSender};
+pub(crate) use self::message_forward_info::{MessageForwardInfo, MessageForwardOrigin};
+use self::secret_chat::SecretChat;
+pub(crate) use self::secret_chat_list::SecretChatList;
+pub(crate) use self::sponsored_message::SponsoredMessage;
+pub(crate) use self::supergroup::Supergroup;
+pub(crate) use self::supergroup_list::SupergroupList;
+pub(crate) use self::user::{BoxedUserStatus, BoxedUserType, User};
+pub(crate) use self::user_list::UserList;
