+++ src/tdlib/secret_chat_list.rs
@@ -4,40 +4,40 @@ use gtk::{gio, glib};
 use indexmap::map::Entry;
 use tdlib::enums::Update;
 
-use crate::session::User;
+use crate::tdlib::SecretChat;
 use crate::Session;
 
 mod imp {
     use super::*;
+    use glib::WeakRef;
     use indexmap::IndexMap;
-    use once_cell::sync::{Lazy, OnceCell};
+    use once_cell::sync::Lazy;
     use std::cell::RefCell;
 
     #[derive(Debug, Default)]
-    pub(crate) struct UserList {
-        pub(super) list: RefCell<IndexMap<i64, User>>,
-        pub(super) session: OnceCell<Session>,
+    pub(crate) struct SecretChatList {
+        pub(super) list: RefCell<IndexMap<i32, SecretChat>>,
+        pub(super) session: WeakRef<Session>,
     }
 
     #[glib::object_subclass]
-    impl ObjectSubclass for UserList {
-        const NAME: &'static str = "UserList";
-        type Type = super::UserList;
+    impl ObjectSubclass for SecretChatList {
+        const NAME: &'static str = "SecretChatList";
+        type Type = super::SecretChatList;
         type Interfaces = (gio::ListModel,);
     }
 
-    impl ObjectImpl for UserList {
+    impl ObjectImpl for SecretChatList {
         fn properties() -> &'static [glib::ParamSpec] {
             static PROPERTIES: Lazy<Vec<glib::ParamSpec>> = Lazy::new(|| {
                 vec![glib::ParamSpecObject::new(
                     "session",
                     "Session",
-                    "The session",
+                    "The session relative to this list",
                     Session::static_type(),
                     glib::ParamFlags::READWRITE | glib::ParamFlags::CONSTRUCT_ONLY,
                 )]
             });
-
             PROPERTIES.as_ref()
         }
 
@@ -49,8 +49,7 @@ mod imp {
             pspec: &glib::ParamSpec,
         ) {
             match pspec.name() {
-                "session" => self.session.set(value.get().unwrap()).unwrap(),
-
+                "session" => self.session.set(Some(&value.get().unwrap())),
                 _ => unimplemented!(),
             }
         }
@@ -63,9 +62,9 @@ mod imp {
         }
     }
 
-    impl ListModelImpl for UserList {
+    impl ListModelImpl for SecretChatList {
         fn item_type(&self, _list_model: &Self::Type) -> glib::Type {
-            User::static_type()
+            SecretChat::static_type()
         }
 
         fn n_items(&self, _list_model: &Self::Type) -> u32 {
@@ -76,62 +75,56 @@ mod imp {
             self.list
                 .borrow()
                 .get_index(position as usize)
-                .map(|(_, u)| u.upcast_ref())
+                .map(|(_, i)| i.upcast_ref())
                 .cloned()
         }
     }
 }
 
 glib::wrapper! {
-    pub(crate) struct UserList(ObjectSubclass<imp::UserList>)
+    pub(crate) struct SecretChatList(ObjectSubclass<imp::SecretChatList>)
         @implements gio::ListModel;
 }
 
-impl UserList {
+impl SecretChatList {
     pub(crate) fn new(session: &Session) -> Self {
-        glib::Object::new(&[("session", session)]).expect("Failed to create UserList")
+        glib::Object::new(&[("session", session)]).expect("Failed to create SecretChatList")
     }
 
-    /// Return the `User` of the specified `id`. Panics if the user is not present.
+    /// Return the `SecretChat` of the specified `id`. Panics if the secret chat is not present.
     /// Note that TDLib guarantees that types are always returned before their ids,
     /// so if you use an `id` returned by TDLib, it should be expected that the
-    /// relative `User` exists in the list.
-    pub(crate) fn get(&self, id: i64) -> User {
-        self.try_get(id).expect("Failed to get expected User")
+    /// relative `SecretChat` exists in the list.
+    pub(crate) fn get(&self, id: i32) -> SecretChat {
+        self.imp()
+            .list
+            .borrow()
+            .get(&id)
+            .expect("Failed to get expected SecretChat")
+            .to_owned()
     }
 
-    pub(crate) fn try_get(&self, id: i64) -> Option<User> {
-        self.imp().list.borrow().get(&id).cloned()
-    }
+    pub(crate) fn handle_update(&self, update: &Update) {
+        if let Update::SecretChat(data) = update {
+            let mut list = self.imp().list.borrow_mut();
 
-    pub(crate) fn handle_update(&self, update: Update) {
-        match update {
-            Update::User(data) => {
-                let mut list = self.imp().list.borrow_mut();
+            match list.entry(data.secret_chat.id) {
+                Entry::Occupied(entry) => entry.get().handle_update(update),
+                Entry::Vacant(entry) => {
+                    let user = self.session().user_list().get(data.secret_chat.user_id);
+                    let secret_chat = SecretChat::from_td_object(&data.secret_chat, &user);
+                    entry.insert(secret_chat);
 
-                match list.entry(data.user.id) {
-                    Entry::Occupied(entry) => entry.get().handle_update(Update::User(data)),
-                    Entry::Vacant(entry) => {
-                        let user = User::from_td_object(data.user, self.session());
-                        entry.insert(user);
+                    let position = (list.len() - 1) as u32;
+                    drop(list);
 
-                        drop(list);
-                        self.item_added();
-                    }
+                    self.items_changed(position, 0, 1);
                 }
             }
-            Update::UserStatus(ref data) => self.get(data.user_id).handle_update(update),
-            _ => {}
         }
     }
 
-    fn item_added(&self) {
-        let list = self.imp().list.borrow();
-        let position = list.len() - 1;
-        self.items_changed(position as u32, 0, 1);
-    }
-
-    pub(crate) fn session(&self) -> &Session {
-        self.imp().session.get().unwrap()
+    pub(crate) fn session(&self) -> Session {
+        self.imp().session.upgrade().unwrap()
     }
 }
